ONSP-518 Configure CoAP separate mode timeout

+++ device-client/coap/src/main/java/org/mbed/coap/server/CoapServer.java
@@ -251,7 +251,7 @@ public class CoapServer extends CoapServerAbstract implements Closeable {
      *
      * @param delayedTransactionTimeout timeout in milliseconds
      */
-    public void setDelayedTransactionTimeout(int delayedTransactionTimeout) {
+    public void setDelayedTransactionTimeout(long delayedTransactionTimeout) {
         this.delayedTransactionTimeout = delayedTransactionTimeout;
     }
